Admin Controller exception Handling Done

+++ mHealthSystem/src/main/java/com/cg/mHealthSystem/Controllers/AdminController.java
@@ -18,6 +18,7 @@ import com.cg.mHealthSystem.Repository.PatientDetailsRepository;
 import com.cg.mHealthSystem.entity.Department;
 import com.cg.mHealthSystem.entity.Doctor;
 import com.cg.mHealthSystem.entity.Nurse;
+import com.cg.mHealthSystem.entity.PatientDetails;
 import com.cg.mHealthSystem.services.AdminService;
 
 @RestController
@@ -65,7 +66,12 @@ public class AdminController {
 	@DeleteMapping("/deleteDoctor/{doctorId}")
 	public boolean deleteDoctor(@PathVariable Integer doctorId)
 	{
-		return adminservice.removeDoctorById(doctorId);
+		Optional<Doctor> doctor = doctorDao.findById(doctorId);
+		if(doctor.isPresent())
+		{
+			return adminservice.removeDoctorById(doctorId);
+		}
+		throw new ResourceNotFoundException();
 	}
 	
 	@PostMapping("/addDepartment")
@@ -89,7 +95,14 @@ public class AdminController {
 	@DeleteMapping("/deletePatient/{patientId}")
 	public boolean deletePatient(@PathVariable Integer patientId)
 	{
-		return adminservice.removePatient(patientId);
+		Optional<PatientDetails> patient =  patientDao.findById(patientId);
+		if(patient.isPresent())
+		{
+			return adminservice.removePatient(patientId);
+		}
+		
+		throw new ResourceNotFoundException();
+		
 	}
